+++ core/src/main/java/com/amplifyframework/auth/result/step/AuthNextResetPasswordStep.java
@@ -117,7 +117,7 @@ public final class AuthNextResetPasswordStep {
      */
     @Override
     public String toString() {
-        return "{" +
+        return "AuthNextResetPasswordStep{" +
                 "resetPasswordStep=" + getResetPasswordStep() +
                 ", additionalInfo=" + getAdditionalInfo() +
                 ", codeDeliveryDetails=" + getCodeDeliveryDetails() +
