+++ core/src/main/java/com/amplifyframework/auth/result/step/AuthResetPasswordStep.java
@@ -33,5 +33,5 @@ public enum AuthResetPasswordStep {
     /**
      * The flow is completed and no further steps are needed.
      */
-    DONE
+    DONE;
 }
