+++ core/src/main/java/com/amplifyframework/auth/result/step/AuthSignUpStep.java
@@ -28,5 +28,5 @@ public enum AuthSignUpStep {
     /**
      * The flow is completed and no further steps are needed.
      */
-    DONE
+    DONE;
 }
